+++ lib/active_interaction.rb
@@ -40,8 +40,9 @@ require 'active_interaction/base'
 
 require 'active_interaction/backports'
 
-I18n.load_path << File.expand_path(
-  File.join(%w[active_interaction locale en.yml]), File.dirname(__FILE__))
+I18n.load_path += Dir[File.expand_path(
+  File.join(%w[active_interaction locale *.yml]), File.dirname(__FILE__))]
+I18n.default_locale = :en
 
 # Manage application specific business logic.
 #
@@ -50,5 +51,5 @@ I18n.load_path << File.expand_path(
 #
 # @since 1.0.0
 #
-# @version 1.2.1
+# @version 1.2.3
 module ActiveInteraction end