+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # [Master][]
 
 - Add an interface filter.
+
+# [1.2.3][] (2014-05-12)
+
+- Fix a bug that raised `ActiveRecord::Rollback` when composing even when not
+  in a transaction.
+
+# [1.2.2][] (2014-05-07)
+
+- Fix a bug that raised `NameError`s when there were invalid nested hash
+  errors.
 - Add missing translation for symbol filters.
 
 # [1.2.1][] (2014-05-02)
@@ -202,7 +212,9 @@
 
 - Initial release.
 
-  [master]: https://github.com/orgsync/active_interaction/compare/v1.2.1...master
+  [master]: https://github.com/orgsync/active_interaction/compare/v1.2.3...master
+  [1.2.3]: https://github.com/orgsync/active_interaction/compare/v1.2.2...v1.2.3
+  [1.2.2]: https://github.com/orgsync/active_interaction/compare/v1.2.1...v1.2.2
   [1.2.1]: https://github.com/orgsync/active_interaction/compare/v1.2.0...v1.2.1
   [1.2.0]: https://github.com/orgsync/active_interaction/compare/v1.1.7...v1.2.0
   [1.1.7]: https://github.com/orgsync/active_interaction/compare/v1.1.6...v1.1.7